### 2022-03-04 #506 @grwhumphries
Presubmission inquiry - stochLAB package for modelling collision risk at offshore wind farms
Submitting Author Grant Humphries
**Reviewers:**
**Editor:**
Language Code English
Submitting Author: Grant Humphries (@grwhumphries)
Other Package Authors: Bruno Caneco (@bcaneco)
Repository:  https://github.com/HiDef-Aerial-Surveying/stochLAB
Submission type: Pre-submission

---

```
Package: stochLAB
Title: Stochastic Collision Risk Model
Version: 1.0.0
    person("Grant", "Humphries", , "grant.humphries@hidefsurveying.co.uk", role = c("aut", "cre")),
    person("Bruno", "Caneco", , "bruno@dmpstats.com", role = "aut"),
    person("Marine Scotland", role = c("fnd", "cph")),
    person("HiDef", role = "cph"),
    person("DMPstats", role = "cph")
  )
Description: Tool for running Collision Risk Models for seabirds on
    offshore wind farms.
License: GPL (>= 3) + file LICENSE
    cli,
    foreach,
    glue,
    logr,
    msm,
    pracma,
    tidyr
    covr
Language: en-GB
LazyDataCompression: bzip2
    https://hidef-aerial-surveying.github.io/stochLAB/
```

	- [ ] text data

- Explain how and why the package falls under these categories (briefly, 1-2 sentences).  Please note any areas you are unsure of: This package is for applying a deterministic model to calculate potential collision risk of seabirds at offshore windfarms.

- If submitting a statistical package, have you already [incorporated documentation of standards into your code via the **srr** package](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr)? No

-   Who is the target audience and what are scientific applications of this package?  This package is for anyone working in the offshore wind industry who is looking to quantify collision risk of seabirds depending on the parameters of the wind farm area or turbines.

-   Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [our criteria for best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)? No

-   (If applicable) Does your package comply with our [guidance around _Ethics, Data Privacy and Human Subjects Research_](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)? N/A
