+++ R/rtnorm_dmp.R
@@ -5,22 +5,7 @@
 #'
 #'
 #' @examples
-#' \donttest{
-#' ## Example case one:  mean < lower while sd = 0
-#' ##
-#' ## original function freezes the R session
-#' msm::rtnorm(n = 1, mean = 1, sd = 0, lower = 0)
-#' ## this version returns NA and issues a warning
-#' rtnorm_dmp(n = 1, mean = 1, sd = 0, lower = 0)
-#'
-#'
-#' ## Example case two: mean = lower while sd = 0
-#' ##
-#' ## original function returns NA with no warning
-#' msm::rtnorm(n = 1, mean = 1, sd = 0, lower = 1)
-#' ## this version returns NA and issues a warning
-#' rtnorm_dmp(n = 1, mean = 1, sd = 0, lower = 1)
-#' }
+#' rtnorm_dmp(n=10,mean=0.4,sd=0.2)
 #'
 #'
 #' @param n An integer value. The number of samples to generate
