### 2022-03-11 #506 @jooolia
Dear @grwhumphries,
Thank you for your pre-submission. This is a statistical package submission and you have replied "no" to the documentation of standards in your code, is this something you intend to do before a full submission or do you need support with this point?
Thanks, Julia
