### 2022-03-18 #506 @jooolia
Yes it would be with the srr package.

I was having a look at the Readme.Rmd and the vignette to get a feel for the outputs and struggled to use them as most of the code snippets are not R chunks and when I change them to R chunks I get several errors. I will try to have a look using some of the examples provided in other files, but it would be nice to clarify why this is not working.

Thanks, Julia
